Add deduplication of events.
Some people put their events in more than one calendar,
and this means that they'll show up twice in our calendar. Sadly, they
do not tend to put in the exact same record in multiple calendars; indeed,
they're often not even consistent about start and end times. So, we
consider two events to be the same thing if they have the same name *and*
they overlap in time.

We then remove the "extra" one -- we check description and location and
throw away the shorter one. We check location first, because descriptions
are likely to vary only in small parts, where locations are often
a useful one ("53 The Glebe, Orpington, G1R 0AA") and a crap one ("G1R 0AA")
and so we throw away the event with the crap one.

Note that this includes moment-range and so needs an npm install.